Refine orchestrator helper usage
