+++ src/diaremot/pipeline/__init__.py
@@ -3,7 +3,12 @@
 __all__ = [
     "audio_pipeline_core",
     "audio_preprocessing",
+    "cli_entry",
     "cpu_optimized_diarizer",
+    "config",
+    "logging_utils",
+    "orchestrator",
+    "outputs",
     "pipeline_checkpoint_system",
     "pipeline_diagnostic",
     "pipeline_healthcheck",
